[CELEBORN-1472] Reduce CongestionController#produceBytes call times.
